[3.2] Move PopupWindow logic to GodotEditText on Android
